requirements.txt
astroid==2.0.1
autopep8==1.3.5
click==6.7
colorama==0.3.9
Flask==1.0.2
Flask-Classful==0.14.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.0.1
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
